+++ ECUAL/ULTRA_SONIC/US.h
@@ -20,6 +20,6 @@ typedef struct{
 
 uint8_t US_u8Init(USElement_t* USDefinition);
 uint8_t US_u8Trigger(USElement_t* USDefinition);
-uint8_t US_u8GetDistance(USElement_t* USDefinition,uint16_t* u8Distance);
+uint8_t US_u8GetDistance(USElement_t* USDefinition,uint8_t* u8Distance);
 
 #endif /* ECUAL_ULTRA_SONIC_US_H_ */
